+++ iam/iam-policy-for-secrets/variables.tf
@@ -1,0 +1,12 @@
+variable "project_access" {
+  type    = map(any)
+  default = {}
+}
+
+variable "region" {
+  type = string
+}
+
+variable "username" {
+  type = string
+}